First Commit
This is a first commit and upload of all lines of code required to allow the program to run.  This was a program designed for a school homework assignment using Python. The purpose of this project was to develop code to calculate the area of a rectangle, triangle, and circle using user given inputs.